fix: Make verbose output print immediately & add 'assertFileContains'.
For the dotline printing, the output is captured and printed after the
test is done, which was the same implementation for verbose mode.
However, I found that sometimes when you make a mistake (like having a
file source another which then unexpectedly already sourced the first)
the process might hang indefinitely. Stopping with CTRL-C fixes the
loop, but any logging or error output will not be printed because of the
capturing. This change now has the output printed directly under verbose
mode, which is what you usually try when something goes wrong anyway.

Also I found I had to check the content of a file quite often, which is
quite an unreadable test.