fix: teardown not call after 'exit 1'
To be able to use environment variables from the test functions, the
test functions are not called in a subshell, but if a test exists
prematurely, the test should exit, but the teardown should still be
called.
This broke because the test that should have catched this did not run as
it was overriden by another test with the same name. Stupid mistake.

The problem is solved by creating an exit trap.

fixes #7