extended matchers in main script + minor output improvement
As there appears to rarely be a use for more than the current small set of
matchers from the extended_matchers script, I've decided to inline it in
the main script for easyer usage.

Reverts the basic matchers to say "expected: a, got: b" again, which is
more understandable than the "expected: a, to equal: b".