reverse forcing shell flags on clients + update docs
In the previous version, I inteded to be compatible to different
restrictive shell flags by adding them in the script. However, this also
forced the flags upon all clients. In this commit I've moved the flags
to the test script, thereby forcing the flags for the runTests.sh script
when testing, but not for clients that use the script. This restores
backwards compatibility.